Begin implementation of libJerryCAN
libJerryCAN has both device and host side components which share a
common types header (jerrycan_types.h) that defines all of the CAN types
used in the system.

On the device side, jerrycan implements the handling and sending of CAN
messages related to each peripheral that is implemented.

On the host side, libjerrycan is a c++ library and a python wrapper
(pyjerrycan) that interacts with SocketCAN on the linux host.

In both cases, the library provides a framework for easily sending and
receiving data over the CAN network, and should be extensible to any
future modules that may need to be hooked into the system.