Support firmware updates over CAN
Add support for flashing firmware via the CAN interface
This adds an additional jerrycan module to handle processing bootloader
specific messages for loading firmware images into the external flash.
Additional subcommands are used for retreiving the currently running
firmware version, restarting the module, and locking in the new image
after a successful flash.

A host side tool was added, jerrycan_updater, to handle sending the
firmware image to the module.