Refactor the ADC TMC2209 register read/write types
In order to prevent any possible endianess issues, reorder all of the
bits in the register definitions to be little endian.  However, because
the UART sends these big endian, need to refactor a bunch of the support
functions to better handle conversions.

While a big change, I think this is cleaner and will be less prone to
errors when attempting to debug and tweak TMC2209 settings.